// src/integrations/supabase/types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.1"
  }
  public: {
    Tables: {
      profiles: {
        Row: {
          avatar_url: string | null
          created_at: string
          display_name: string | null
          id: string
          updated_at: string
        }
        Insert: {
          avatar_url?: string | null
          created_at?: string
          display_name?: string | null
          id: string
          updated_at?: string
        }
        Update: {
          avatar_url?: string | null
          created_at?: string
          display_name?: string | null
          id?: string
          updated_at?: string
        }
        Relationships: []
      }
      properties: {
        Row: {
          address: string
          city: string
          client_name: string
          created_at: string
          email_1: string | null
          email_2: string | null
          email_3: string | null
          first_name: string
          id: number
          last_name: string
          last_seen_1: string | null
          last_seen_2: string | null
          last_seen_3: string | null
          mailing_address: string | null
          mailing_city: string | null
          mailing_state: string | null
          mailing_zip: string | null
          phone_1: string | null
          phone_2: string | null
          phone_3: string | null
          state: string
          type_1: string | null
          type_2: string | null
          type_3: string | null
          wrong_1: boolean | null
          wrong_2: boolean | null
          wrong_3: boolean | null
          zipcode: string
        }
        Insert: {
          address: string
          city: string
          client_name: string
          created_at?: string
          email_1?: string | null
          email_2?: string | null
          email_3?: string | null
          first_name: string
          id?: number
          last_name: string
          last_seen_1?: string | null
          last_seen_2?: string | null
          last_seen_3?: string | null
          mailing_address?: string | null
          mailing_city?: string | null
          mailing_state?: string | null
          mailing_zip?: string | null
          phone_1?: string | null
          phone_2?: string | null
          phone_3?: string | null
          state: string
          type_1?: string | null
          type_2?: string | null
          type_3?: string | null
          wrong_1?: boolean | null
          wrong_2?: boolean | null
          wrong_3?: boolean | null
          zipcode: string
        }
        Update: {
          address?: string
          city?: string
          client_name?: string
          created_at?: string
          email_1?: string | null
          email_2?: string | null
          email_3?: string | null
          first_name?: string
          id?: number
          last_name?: string
          last_seen_1?: string | null
          last_seen_2?: string | null
          last_seen_3?: string | null
          mailing_address?: string | null
          mailing_city?: string | null
          mailing_state?: string | null
          mailing_zip?: string | null
          phone_1?: string | null
          phone_2?: string | null
          phone_3?: string | null
          state?: string
          type_1?: string | null
          type_2?: string | null
          type_3?: string | null
          wrong_1?: boolean | null
          wrong_2?: boolean | null
          wrong_3?: boolean | null
          zipcode?: string
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof DefaultSchema["CompositeTypes"]
    | { schema: keyof DatabaseWithoutInternals },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof DefaultSchema["CompositeTypes"]
    ? DefaultSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never